Refactor auth domain structure
